Refactor: Update router implementation with controller on notification service and gateway
