+++ services/notification-service/src/routes/health.route.ts
@@ -1,13 +1,9 @@
-import { Request, Response } from "express";
-import { catchErrors } from "@jeffreybernadas/service-hub-helper";
-import { log } from "@notifications/utils/logger.util";
+import { Router } from "express";
+import healthCheckHandler from "@notifications/controller/health.controller";
 
-const healthCheckHandler = catchErrors(async (_req: Request, res: Response) => {
-  log.info("Health check endpoint called");
+const healthCheckRouter = Router();
 
-  res.status(200).json({
-    status: "Notification Service is healthy.",
-  });
-});
+healthCheckRouter.get("/", healthCheckHandler);
 
-export default healthCheckHandler;
+export default healthCheckRouter;
+
